Check dimensions of the parameters

check_dimensions checks dimensions of the parameters. If wrong, throw an error.

Usage

check_dimensions(
  p,
  root.state,
  shifts,
  variance,
  selection.strength = NULL,
  optimal.value = NULL
)

Arguments

p

dimension of the trait simulated

root.state

(list) state of the root, with: random : random state (TRUE) or deterministic state (FALSE) value.root : if deterministic, value of the character at the root exp.root : if random, expectation of the character at the root var.root : if random, variance of the character at the root

shifts

(list) position and values of the shifts : edges : vector of the K id of edges where the shifts are values : matrix p x K of values of the shifts on the edges (one column = one shift) relativeTimes : vector of dimension K of relative time of the shift from the parent node of edges

variance

variance-covariance matrix size p x p

selection.strength

matrix of selection strength size p x p (OU)

optimal.value

vector of p optimal values at the root (OU)

Value

Nothing