Squared Mahalanobis Distance

compute_mahalanobis_distance.simple computes the squared Mahalanobis distance between the data and mean at tips of the data in the simple case where the inverse of the variance matrix is given.

compute_mahalanobis_distance.simple(phylo, Y_data_vec, sim, Sigma_YY_chol_inv,
  miss = rep(FALSE, dim(sim)[1] * length(phylo$tip.label)), ...)

Arguments

phylo
Input tree.

sim
(list) : result of function simulate.

Y_data
: vector indicating the data at the tips.

Sigma_YY_inv
: invert of the variance-covariance matrix of the data.

Value

squared Mahalanobis distance between data and mean at the tips.

Details

This function takes parameters sim and Sigma_YY_inv from compute_mean_variance.simple. It uses function extract_simulate_internal to extract the needed quantities from these objects.