Compute variances of differences between nodes and parents.

compute_var_diff computes variances of differences between all the nodes and their parents.

compute_var_diff.BM(phylo, conditional_law_X)

Arguments

phylo

a phylogenetic tree

conditional_law_X

result of function compute_E

Value

array p x p x Nedge containing, for each edge e finishing at node i, the quantity Var[Z_i-Z_pa(i)|Y].