Expectation conditional to the tips of the completed log-likelihood.

conditional_expectation_log_likelihood.OU computes the expectation conditional to the tips of the completed log-likelihood, given the first and second order moments of the nodes, and the parameters of the OU process.

conditional_expectation_log_likelihood.OU.OLD(phylo, conditional_law_X, sigma2,
  mu, alpha)

Arguments

phylo

Input tree.

conditional_law_X

result of function compute_E.OU, containing first and second moments.

sigma2

variance of the OU.

mu

mean of the root.

alpha

selection strength of the OU.

Value

double : value of the function

Details

This function uses functions compute_var_diff.OU and compute_diff_exp.OU in the process. Carefull : only works if the root is stationary, and shifts at nodes.