Get Model Selection Criterion

This function takes an object of class PhyloEM, result of function PhyloEM, and return the values of the model selection criterion for each value of K.

Usage

get_criterion(res, method.selection = NULL)

Arguments

res

an object of class PhyloEM, result of function PhyloEM.

method.selection

select the parameters to plot. One of "LINselect", "DDSE", "Djump" or "likelihood" (for un-penalized likelihood). Default to "LINselect". See params_process.PhyloEM.

Value

A named vector with the values of the criterion for each number of shift K.

See also