Ancestral State Reconstruction

imputed_traits.PhyloEM takes an object of class PhyloEM, and returns the imputed traits values, either at the internal nodes (ancestral state reconstruction) or at the tips (data imputation)

imputed_traits(x, ...)

# S3 method for PhyloEM
imputed_traits(
  x,
  trait = 1,
  save_all = FALSE,
  where = c("nodes", "tips"),
  what = c("imputed", "variances", "expectations"),
  params = NULL,
  method.selection = NULL,
  reconstructed_states = NULL,
  ...
)

Arguments

x

an object of class PhyloEM.

...

further arguments to be passed on to params_process.PhyloEM

trait

an integer giving the trait to extract. Default to 1.

save_all

if TRUE, arguments where and what are ignored, and all the moments are kept for further extraction with the same function, specifying the argument reconstructed_states. Default to FALSE.

where

either "nodes" for ancestral state reconstruction, or "tips" for data imputation.

what

the quantity to retrieve. Either the imputed traits (default), their conditional variances, or the simple expectations under the selected process.

params

(optional) some user-specified parameters. Must be of class params_process. If left blank, they are extracted using the method.selection argument (see below).

method.selection

(optional) the method selection to be used. One of "LINselect", "DDSE", "Djump". Default to "LINselect".

reconstructed_states

if the reconstructed states have already been computed (by a previous call of the function, with save_all=TRUE), they can be passed on here (avoids multiple computations of the E step).

Value

A matrix or array with the computed quantities.

Methods (by class)

PhyloEM: PhyloEM object

See also