Compute the actualizations factors to apply to the incidence matrix.

incidence_matrix_actualization_factors computes a ntaxa x nedges matrix of the (1 - exp(-alpha * (t_i - t_pa(j) - nu_j * l_j)))_i tip, j node. This matrix is to be multiplied to the incidence matrix with an outer product.

incidence_matrix_actualization_factors(tree, selection.strength,
  relativeTimes_tree = 0, times_shared = compute_times_ca(tree))

Arguments

tree
a phylogenetic tree.

selection.strength
the selection strength of the process.

relativeTimes_tree
a nedge vector of relative times associated with the branches.

times_shared
a matrix, result of function compute_times_ca.

Value

Matrix of size ntaxa x nedges