Initialization for incidence matrix

init.incidence.matrix initialize the matrix updated in update.incidence.matrix for the computation of the incidence matrix in incidence.matrix.

init.incidence.matrix(phy)

Arguments

phy
Input tree.

Value

Matrix with nNodes rows and ntaxa column.

Details

The initialized matrix has ntaxa column and nNodes rows. Each node represent its parental branch. A row corresponding to a tip i is initialized to a vector of zeros, with only entry i equal to one. (Branch ending at tip i is only in the i^th lineage)