Model Selection of a fitted object

model_selection does the model selection on a fitted PhyloEM object, and returns the same fitted object.

model_selection(x, ...)

# S3 method for PhyloEM
model_selection(x, method.selection = c("LINselect", "DDSE",
  "Djump"), C.BM1 = 0.1, C.BM2 = 2.5, C.LINselect = 1.1,
  independent = FALSE, ...)

Arguments

x

a fitted PhyloEM object

...

Further arguments to be passed to estimateEM, including tolerance parameters for stopping criterions, maximal number of iterations, etc.

method.selection

Method selection to be used. Several ones can be used at the same time. One of "LINselect" for the Baraud Giraud Huet LINselect method; "DDSE" for the Slope Heuristic or "Djump" for the Jump Heuristic, last two based the Birgé Massart method.

C.BM1

Multiplying constant to be used for the BigeMassart1 method. Need to be positive. Default to 0.1.

C.BM2

Multiplying constant to be used for the BigeMassart2 method. Default to 2.5.

C.LINselect

Multiplying constant to be used for the LINselect method. Need to be greater than 1. Default to 1.1.

independent

Are the trait assumed to be independent from one another? Default to FALSE. OU in a multivariate setting only works if TRUE.

Value

The same object, but with a slot corresponding to the model selection used. See function params_process.PhyloEM to retrieve the selected parameters.

Methods (by class)

PhyloEM: PhyloEM object

See also

PhyloEM, params_process.PhyloEM, imputed_traits.PhyloEM