Parameter estimates

params takes an object of class PhyloEM, and returns the inferred parameters of the process.

# S3 method for PhyloEM
params_process(
  x,
  method.selection = NULL,
  K = NULL,
  alpha = NULL,
  rBM = FALSE,
  init = FALSE,
  ...
)

Arguments

x

an object of class PhyloEM

method.selection

(optional) the method selection to be used. One of "LINselect", "DDSE", "Djump". Default to "LINselect".

K

(optional) an integer giving the number of shifts for which to retrieve the parameters. Default to NULL (automatically selected number of shifts, see method.selection argument).

alpha

(optional) a value of alpha for which to retrieve the parameters. Can be an (un-ambiguous) estimation of the true value. If specified, then K must be precised too. Default to NULL (automatically selected value, see method.selection argument).

rBM

(optional) if TRUE, and if the process is "scOU", returns the raw parameters of the BM on the re-scaled tree. Default to FALSE, except if the selection strength is negative (see doc of PhyloEM for an explanation of this particular case).

init

(optional) if TRUE, gives the parameters from the initialization of the EM. Default to FALSE. This has no effect if K is not specified.

...

unused.

Value

An object of class params_process.

See also