Plot all the equivalent solutions.

plot.enumerate_parsimony plots a representation of all the equivalent solutions.

Usage

# S3 method for class 'enumerate_parsimony'
plot(x, numbering = FALSE, nbr_col = 3, ...)

Arguments

x

an object of class enumerate_parsimony, result of function enumerate_parsimony

numbering

whether to number the solutions. Default to FALSE.

nbr_col

the number of columns on which to display the plot. Default to 3.

...

further arguments to be passed to plot.phylo or nodelabels.

Value

A plot of the equivalent shifts allocations.

Details

This function uses functions plot.phylo and nodelabels for the actual plotting of the trees.

See also