Plot for class simul_process

This function takes an object of class params_process, and plots them along with some data at the tips of the tree.

# S3 method for params_process
plot(x, phylo, data = NULL, traits,
  automatic_colors = TRUE, color_characters = "black",
  color_edges = "black", plot_ancestral_states = FALSE,
  ancestral_states = NULL, imposed_scale, ancestral_cex = 2,
  ancestral_pch = 19, value_in_box = FALSE, ancestral_as_shift = FALSE,
  shifts_cex = 0.6, shifts_bg = "chocolate4", root_bg = "chocolate4",
  shifts_adj = 0, root_adj = 1, color_shifts_regimes = FALSE,
  regime_boxes = FALSE, alpha_border = 70, show.tip.label = FALSE,
  label_cex = 0.5, label_offset = 0, axis_cex = 0.7, edge.width = 1,
  margin_plot = NULL, gray_scale = FALSE, ...)

Arguments

x
an object of class params_process.

phylo
a phylogenetic tree.

data
a matrix of data at the tips of the tree. Must have p rows and ntaxa columns. If these are simulated, use the extract.simul_process function.

traits
a vector of integers giving the numbers of the trait to be plotted. Default to 1:p (all the traits).

automatic_colors
whether to color the edges automatically according to their regimes. Default to TRUE. If FALSE, colors can be mannually precised through arguments color_characters and colro_edges (see below).

color_characters
if automatic_colors=FALSE, a vector of colors for the tips of the tree.

color_edges
if automatic_colors=FALSE, a vector of colors for the edges of the tree.

plot_ancestral_states
whether to plot the ancestral traits inffered at the internal nodes of the tree. Only available if only one trait is plotted. Default to FALSE.

ancestral_states
if plot_ancestral_states=TRUE, the ancestral states must be specified. If these are simulated, use the extract.simul_process function.

imposed_scale
if plot_ancestral_states=TRUE, a vector precising the imposed scale for the ancestral states plotting. Useful to make comparisons. Default to the plotted trait.

ancestral_cex
if plot_ancestral_states=TRUE, the size of the ancestral states on the tree. Default to 2.

ancestral_pch
if plot_ancestral_states=TRUE, the symbol used of the ancestral states. Default to circles (pch=19).

value_in_box
whether to plot the value of the shift in a box on the edges. Only available when only one trait is plotted. Can be difficult to read on big trees. The size of the text in the boxes is controlled by parameter. Default to FALSE.

ancestral_as_shift
whether to represent the ancestral value at the root as an ancestral shift on the root edge. Default to FALSE. shifts_cex (see below).

shifts_cex
if value_in_box=TRUE, the size of the text in the boxes. Default to 0.8.

shifts_bg
if value_in_box=TRUE, the background color of the boxes.

root_bg
if value_in_box=TRUE and ancestral_as_shift=TRUE, the background color of the ancestral boxe.

shifts_adj
the adj parameter for the shifts position on the edges. Default to 0 (begining of the edge).

root_adj
if ancestral_as_shift=TRUE, the adj parameter for the ancestral value position on the root edge. Default to 1.

color_shifts_regimes
whether to color each shift according to its regime (default to the same color of the edge it's on). Default to FALSE.

regime_boxes
whether to draw a box showing all the tips below a given. The transparency of the border of the box is controlled by parameter alpha_border (see below).

alpha_border
if regime_boxes=TRUE, the alpha parameter of the border of the box. Default to 70.

show.tip.label
whether to show the tip labels. Default to FALSE.

label_cex
if show.tip.label=TRUE, the size of the labels. Default to 0.5.

label_offset
if show.tip.label=TRUE, the size of the offset between the tree and the labels. Default to 0.

axis_cex
cex for the label values of the plot. Default to 0.7.

edge.width
width of the edge. Default to 1.

margin_plot
vector giving the margin to around the plot. Default to c(0, 0, 0, 0).

gray_scale
if TRUE, the colors are replaced by a gray scale. Default to FALSE.

...
further arguments to be passed to plot.phylo.

See also

simul_process, plot.PhyloEM, params_BM, params_OU