exact qr.solve

This is the same function as qr.solve, but it throws an error if an exact fit cannot be found (instead of returning a least square fitted value, which is the default behavior of qr.solve).

qr.solve_exact(a, b, tol = 1e-07)

Arguments

a

a QR decomposition or a rectangular matrix.

b

a vector or matrix of right-hand sides of equations. incidence.matrix.

tol

the tolerance for detecting linear dependencies in the columns of x. Only used if LAPACK is false and x is real.