Update function for incidence matrix

update.incidence.matrix.full updates the matrix initialized in init.incidence.matrix.full for the computation of the incidence matrix in incidence.matrix.full.

update.incidence.matrix.full(daughtersParams, parent, ...)

Arguments

daughtersParams
: rows of updated matrix corresponding to the daughters of the current node.

Value

Vector of length ntaxa + nNodes, indicating to which lineages the branch above the current node belongs to.

Details

A node belongs to all the lineages of its daughters.