Make the result lighter

enlight.PhyloEM takes an object of class PhyloEM, and returns the same object, without saving the quantities that can be easily re-computed using function imputed_traits.PhyloEM.

Usage

enlight(x)

# S3 method for class 'PhyloEM'
enlight(x)

Arguments

x

an object of class PhyloEM.

Value

Same as entry, lighter.

Details

The resulting object can be much lighter, saving a lot of memory space, but each call to the function imputed_traits.PhyloEM will be longer. As function plot.PhyloEM relies on this function, this makes the plotting also longer. This has the same effect as setting the option "light_result=TRUE" in the call of PhyloEM.

Methods (by class)

enlight(PhyloEM): PhyloEM object

See also