Find values given edges. OU stationary case. Ultrametric tree.

equivalent_shifts_values computes the values of the shifts given all the possible allocations computed by function equivalent_shifts_edges.

equivalent_shifts_values(phylo, eq_shifts_edges, T_tree_ac, vect_Y, p)

Arguments

phylo
a phylogenetic tree, class phylo.

eq_shifts_edges
matrix (optional) result of function equivalent_shifts_edges.

T_tree_ac
matrix of incidence of the tree, result of function incidence.matrix, actualized with coefficients computed by function incidence_matrix_actualization_factors.

shifts
a list of positions and values of original shifts.

beta_0
value of the original optimal value at the root.

Value

Named list, with "shifts_values" a matrix of shifts values corresponding to the shifts positions in eq_shifts_edges; and "betas_0" a vector of corresponding root optimal values.

Details

This function uses the linear representation of the problem. It fist compute the mean at the tips given by the original shifts positions and values, and then uses function qr.solve to find back the values of the shifts, given their various positions, and the means at the tips. Function compute_actualization_factors is used to compute the actualization factor that multiplies the shifts values at the tips. Careful, only works for ULTRAMETRIC trees.