Function to estimate alpha

optimize to maximize the conditional expectation log likelihood in alpha. The interval is set to [alpha_old/2, 2*alpha_old], supposing that the previous guess of alpha_old is not far from reality.

Usage

estimate.alpha(
  phylo,
  conditional_law_X,
  sigma2,
  mu,
  shifts,
  alpha_old,
  max_selection.strength
)

Arguments

phylo

Input tree.

conditional_law_X

result of function compute_E.OU

sigma2

variance of params

mu

mean of the root state

shifts

list of shifts on the tree

alpha_old

previous estimation of the selection strength

max_selection.strength

the maximal value of alpha authorized by the user

Value

double : estimation of alpha

Details

This function uses functions compute_var_diff.OU and compute_diff_exp.OU in the process. Careful : only works if the root is stationary, and shifts at nodes.