Plot all the equivalent solutions.

plot.equivalent_shifts plots a representation of all the equivalent shifts allocations, with a representation of the shifts and their values, and a coloration of the branches in term of regimes.

Usage

# S3 method for class 'equivalent_shifts'
plot(
  x,
  trait = 1,
  show_shifts_values = TRUE,
  numbering = FALSE,
  colors_tips = NULL,
  nbr_col = 3,
  gray_scale = FALSE,
  edge.width = 2,
  shifts_cex = 1.2,
  ...
)

Arguments

x

an object of class equivalent_shifts, result of function equivalent_shifts

trait

(integer) the trait to be plotted, if multivariate. Default to 1.

show_shifts_values

whether to show the equivalent shifts values or not. Default to FALSE.

numbering

whether to number the solutions. Default to FALSE.

colors_tips

user-provided colors for the tips of the tree. A vector vector with as many colors as there are tips. Will be automatically computed if not provided.

nbr_col

the number of columns on which to display the plot. Default to 3.

gray_scale

if TRUE, the colors are replaced by a gray scale. Default to FALSE.

edge.width

width of the edge. Default to 1.

shifts_cex

if value_in_box=TRUE, the size of the text in the boxes. Default to 0.8.

...

further arguments to be passed to plot.phylo.

Value

A plot of the equivalent shifts allocations.

Details

This function uses function plot.phylo for the actual plotting of the trees.

See also